Deepening project structure
The SDK was sitting at root level of the project meaning that adding
modules at the same level wasn't possible. I have pushed the SDK into an
"sdk" modules within the project so that I can add in additional modules
like a sample app that allows a debugger to step through the local sdk
internals.