Introducing Sample App
Initial creation of sample app with 2 functions: identify and track. I
will be adding features to this sample app to display all of the
available features in the near future.

This commit also includes some fixes for a push notification bug that
was introduced with the safety improvements added previously. These safe
calls prevented an expected exception from occurring so the small icon
was never correctly set causing the notification to be swallowed by the
system.