Black Box Testing
Added some black box instrumented tests for the SDK to provide a safety
net for future refactoring. Added a test sample app to be used for the
tests. The tests check for appropriate network requests being sent for
using the public methods of the SDK.

Some small refactoring was required internally to allow the tests to be
effective. The difficulties encountered and their solutions are
documented in the test files within the test sample app.

Also included some cleanup of the build.gradle files for consistency
between modules using shared libraries.