Integrating CI
The repo initially had no tests and no CI setup. Now that we have some
instrumented tests we want CI to run these tests whenever a PR is
created. This commit introduces a `travis.yml` file to configure the
travis build, and a `travis.sh` file to do some initial setup and
control of the build steps.