Integrating CI
The repo initially had no tests and no CI setup. Now that we have some
instrumented tests we want CI to run these whenever a PR is created.
This commit introduces a yaml file for configuring the build and a shell
script for setup and control of the build process. I also added a set of
guidelines for any potential contributors