Connect SDK submodule
I added in a new submodule to the project for the ConnectSDK. This new
submodule is a dependency of the existing code so we can use new code
piece by piece and then eventually just depend directly on the new code.

I also renamed the project and all submodules for neatness and clarity.