Flaky Tests Fix
One of the instrumented tests was failing every now and again so I had a
look into fixing it. Turns out there was a lot wrong with the tests due
to the threading causing tests to finish before requests were returning.
I've integrated IdlingResources, CountDownLatches, and am now injecting
a test OkHttpClient to the SDK to alleviate these issues.