UI begone!
So I might have gone a bit too far with this one. My tests don't depend
on UI at all, so I've removed all of it. I also only need to start an
activity to get access to an instance of Application, so I've changed
the setup to launch the activity once and then hold onto the Application
which has led to the tests going from >15 secs to ~2 secs.