More timing issues
Fixed some more timing issues where tests that expected no requests to
be made were passing because the test finished before the request was
actually being made. This was due to disabled config not having enough
time to be stored properly which allowed requests to be made.

There was also an issue with requests that were dependent on identify to
be called first so I added another latch to signify when the extra
request is ready to be made.