Increasing wait time
I think some tests are failing because the emulator on CI is too slow so
I've increased the length of the time that the tests will wait for each
request to be made. The wait period will end once the request is made so
it shouldn't increase the test time.